Renamed CommandExecutor to CommandExecutorError
