+++ src/execution-engines/commands/unknown.rs
@@ -1,13 +1,13 @@
 use std::error::Error;
 
-use crate::exe_engine::cores::{Command, CommandError, CommandExecutor, CommandResult};
+use crate::exe_engine::cores::{Command, CommandExecutorError, CommandExecutor, CommandResult};
 use crate::storage::cores::Storage;
 
 pub struct Unknown;
 
 impl CommandExecutor for Unknown {
     fn execute(_: &mut Box<dyn Storage>, cmd: &Command) -> Result<CommandResult, Box<dyn Error>> {
-        let err = Box::new(CommandError::UnknownCommand(String::from(cmd.name)));
+        let err = Box::new(CommandExecutorError::UnknownCommand(String::from(cmd.name)));
         Err(err)
     }
 }
