+++ src/execution-engines/cores/mod.rs
@@ -1,5 +1,5 @@
 pub use command::*;
-pub use command_errors::*;
+pub use command_executor_errors::*;
 pub use command_executor::*;
 pub use command_parameter::*;
 pub use command_result::*;
@@ -10,6 +10,6 @@ mod command;
 mod command_executor;
 mod command_parameter;
 mod command_result;
-mod command_errors;
+mod command_executor_errors;
 mod engine_errors;
 mod engine;
